Added a seperate Firebase class (FirebaseAccessor)
This firebase class purpose is to interact with the firebase (creating data or fetching data)
Implementation is not final, needs to fetch rider requests
